src/requirements.txt
langchain>=0.3.20,<0.4.0
langchain-core>=0.3.62,<0.4.0
langgraph>=0.2.50,<0.4.0
langchain-openai>=0.2.12,<0.3.0
langchain-google-genai>=2.1.5,<3.0.0
ollama==0.6.1
langchain-ollama>=0.2.0,<0.3.0
traceloop-sdk==0.53.3
yfinance>=0.2.40
vaderSentiment>=3.3.2